requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
google-generativeai==0.8.0
requests==2.31.0
gunicorn==21.2.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-telegram-bot==21.10
nest-asyncio==1.6.0
